third_party/blink/web_tests/platform/generic/fast/dom/frame-loading-via-document-write-expected.txt



--------
Frame: 'topRow'
--------
This page was successfully loaded.
My protocol is file:
My referrer is blank


--------
Frame: 'middleRow'
--------
Image loaded.

--------
Frame: 'bottomRow'
--------
Image loaded.
